[Cleanup] Split even more functions from Mist.cpp/.h
