+++ src/_Plugin_Helper.cpp
@@ -9,6 +9,7 @@
 #include "src/Globals/Settings.h"
 #include "src/Globals/SecuritySettings.h"
 #include "src/Helpers/Misc.h"
+#include "src/Helpers/StringParser.h"
 
 
 PluginTaskData_base *Plugin_task_data[TASKS_MAX] = { nullptr, };
