+++ src/src/Commands/UDP.cpp
@@ -8,6 +8,7 @@
 #include "../Globals/Settings.h"
 #include "../Helpers/Misc.h"
 #include "../Helpers/StringConverter.h"
+#include "../Helpers/StringParser.h"
 
 String Command_UDP_Test(struct EventStruct *event, const char *Line)
 {
