+++ src/src/DataStructs/NodeStruct.h
@@ -13,17 +13,14 @@
 #define NODE_TYPE_ID_ARDUINO_EASY_STD      65
 #define NODE_TYPE_ID_NANO_EASY_STD         81
 
+String getNodeTypeDisplayString(byte nodeType);
 
 /*********************************************************************************************\
 * NodeStruct
 \*********************************************************************************************/
 struct NodeStruct
 {
-  NodeStruct() :
-    build(0), age(0), nodeType(0), webgui_portnumber(0)
-  {
-    for (byte i = 0; i < 4; ++i) { ip[i] = 0; }
-  }
+  NodeStruct();
 
   String    nodeName;
   IPAddress ip;
