+++ src/src/Globals/RamTracker.cpp
@@ -2,7 +2,10 @@
 
 #include "../../ESPEasy_Log.h"
 #include "../../ESPEasy_fdwdecl.h"
-#include "Statistics.h"
+
+#include "../Globals/Statistics.h"
+
+#include "../Helpers/Memory.h"
 #include "../Helpers/Misc.h"
 
 #ifndef BUILD_NO_RAM_TRACKER
