+++ src/src/Helpers/ESPEasy_FactoryDefault.h
@@ -7,5 +7,10 @@
  \*********************************************************************************************/
 void ResetFactory();
 
+/*********************************************************************************************\
+   Collect the stored preference for factory default
+\*********************************************************************************************/
+void applyFactoryDefaultPref();
+
 
 #endif
