+++ src/src/Helpers/Memory.cpp
@@ -1,0 +1,90 @@
+#include "Memory.h"
+
+
+#ifdef ESP8266
+extern "C" {
+#include "user_interface.h"
+}
+#endif
+
+
+/*********************************************************************************************\
+   Memory management
+\*********************************************************************************************/
+
+
+// For keeping track of 'cont' stack
+// See: https://github.com/esp8266/Arduino/issues/2557
+//      https://github.com/esp8266/Arduino/issues/5148#issuecomment-424329183
+//      https://github.com/letscontrolit/ESPEasy/issues/1824
+#ifdef ESP32
+
+// FIXME TD-er: For ESP32 you need to provide the task number, or NULL to get from the calling task.
+uint32_t getCurrentFreeStack() {
+  register uint8_t *sp asm ("a1");
+
+  return sp - pxTaskGetStackStart(NULL);
+}
+
+uint32_t getFreeStackWatermark() {
+  return uxTaskGetStackHighWaterMark(NULL);
+}
+
+// FIXME TD-er: Must check if these functions are also needed for ESP32.
+bool canYield() {
+  return true;
+}
+
+#else // ifdef ESP32
+
+uint32_t getCurrentFreeStack() {
+  // https://github.com/esp8266/Arduino/issues/2557
+  register uint32_t *sp asm ("a1");
+
+  return 4 * (sp - g_pcont->stack);
+}
+
+uint32_t getFreeStackWatermark() {
+  return cont_get_free_stack(g_pcont);
+}
+
+bool canYield() {
+  return cont_can_yield(g_pcont);
+}
+
+bool allocatedOnStack(const void *address) {
+  register uint32_t *sp asm ("a1");
+
+  if (sp < address) { return false; }
+  return g_pcont->stack < address;
+}
+
+#endif // ESP32
+
+
+/********************************************************************************************\
+   Get free system mem
+ \*********************************************************************************************/
+unsigned long FreeMem(void)
+{
+  #if defined(ESP8266)
+  return system_get_free_heap_size();
+  #endif // if defined(ESP8266)
+  #if defined(ESP32)
+  return ESP.getFreeHeap();
+  #endif // if defined(ESP32)
+}
+
+unsigned long getMaxFreeBlock()
+{
+  unsigned long freemem = FreeMem();
+
+  #ifdef CORE_POST_2_5_0
+
+  // computing max free block is a rather extensive operation, so only perform when free memory is already low.
+  if (freemem < 6144) {
+    return ESP.getMaxFreeBlockSize();
+  }
+  #endif // ifdef CORE_POST_2_5_0
+  return freemem;
+}
