+++ src/src/Helpers/PeriodicalActions.cpp
@@ -18,7 +18,9 @@
 #include "../Globals/Statistics.h"
 #include "../Helpers/ESPEasyRTC.h"
 #include "../Helpers/Hardware.h"
+#include "../Helpers/Memory.h"
 #include "../Helpers/Misc.h"
+#include "../Helpers/StringGenerator_System.h"
 
 
 /*********************************************************************************************\
