+++ src/src/Helpers/SystemVariables.cpp
@@ -1,5 +1,6 @@
 #include "SystemVariables.h"
 
+#include "../../ESPEasy_common.h"
 #include "../../ESPEasy-Globals.h"
 
 #include "../DataStructs/TimingStats.h"
